Fixed bug with mAh becomming NULL when not all data was fetched from the modules meanwhile the bypass was running. This caused the loadresistance config to put a null into one of the variables not recovering after.

+++ ESPController/include/defines.h
@@ -197,7 +197,7 @@ struct CellModuleInfo
   //Total PWM time
   uint32_t PWMTime;
   //bleeded mAh calculated
-  double mAh;
+  float mAh;
 };
 
 // This enum holds the states the controller goes through whilst
